requirements.txt
Django==2.2.17
Pillow==8.0.1
mysqlclient==2.0.2
python-decouple==3.3
django-auditlog==0.4.7
